moved TestCollector and ExpectCalls to collector.py
